Refactored click helper methods to own utils module.
