+++ src/WhiteLibrary/keywords/items/listview.py
@@ -1,7 +1,7 @@
 from TestStack.White.UIItems import ListView
 from WhiteLibrary.keywords.librarycomponent import LibraryComponent
 from WhiteLibrary.keywords.robotlibcore import keyword
-from WhiteLibrary.keywords.items.uiitem import UiItemKeywords
+from WhiteLibrary.utils.click import Clicks
 
 class ListViewKeywords(LibraryComponent):
     @keyword
@@ -22,7 +22,7 @@ class ListViewKeywords(LibraryComponent):
         | Double Click Listview Cell | id:addressList | Street | 0 | # double click cell in the column "Street" of the first row |
         """
         cell = self._get_cell(locator, column_name, row_index)
-        UiItemKeywords.double_click(cell, x_offset, y_offset)
+        Clicks.double_click(cell, x_offset, y_offset)
 
     @keyword
     def double_click_listview_cell_by_index(self, locator, row_index, column_index, x_offset=0, y_offset=0):
@@ -42,7 +42,7 @@ class ListViewKeywords(LibraryComponent):
         | Double Click Listview Cell By Index | id:addressList | 0 | 0 |
         """
         cell = self._get_cell_by_index(locator, row_index, column_index)
-        UiItemKeywords.double_click(cell, x_offset, y_offset)
+        Clicks.double_click(cell, x_offset, y_offset)
 
     @keyword
     def double_click_listview_row(self, locator, column_name, cell_text, x_offset=0, y_offset=0):
@@ -61,7 +61,7 @@ class ListViewKeywords(LibraryComponent):
         | Double Click Listview Row | id:addressList | City | Helsinki | # double click row that has the text "Helsinki" in the column "City" |
         """
         row = self._get_row(locator, column_name, cell_text)
-        UiItemKeywords.double_click(row, x_offset, y_offset)
+        Clicks.double_click(row, x_offset, y_offset)
 
     @keyword
     def double_click_listview_row_by_index(self, locator, row_index, x_offset=0, y_offset=0):
@@ -79,7 +79,7 @@ class ListViewKeywords(LibraryComponent):
         | Double Click Listview Row By Index | id:addressList | 4 |
         """
         row = self._get_row_by_index(locator, row_index)
-        UiItemKeywords.double_click(row, x_offset, y_offset)
+        Clicks.double_click(row, x_offset, y_offset)
 
     @keyword
     def get_listview_cell_text(self, locator, column_name, row_index):
@@ -262,7 +262,7 @@ class ListViewKeywords(LibraryComponent):
         See `Double Click Listview Cell` for details about arguments ``locator``, ``column_name``, and ``row_index``.
         """
         cell = self._get_cell(locator, column_name, row_index)
-        UiItemKeywords.right_click(cell, x_offset, y_offset)
+        Clicks.right_click(cell, x_offset, y_offset)
 
     @keyword
     def right_click_listview_cell_by_index(self, locator, row_index, column_index, x_offset=0, y_offset=0):
@@ -271,7 +271,7 @@ class ListViewKeywords(LibraryComponent):
         See `Double Click Listview Cell By Index` for details about arguments ``locator``, ``row_index``, and ``column_index``.
         """
         cell = self._get_cell_by_index(locator, row_index, column_index)
-        UiItemKeywords.right_click(cell, x_offset, y_offset)
+        Clicks.right_click(cell, x_offset, y_offset)
 
     @keyword
     def right_click_listview_row(self, locator, column_name, cell_text, x_offset=0, y_offset=0):
@@ -280,7 +280,7 @@ class ListViewKeywords(LibraryComponent):
         See `Double Click Listview Row` for details about the arguments ``locator``, ``column_name``, and ``cell_text``.
         """
         row = self._get_row(locator, column_name, cell_text)
-        UiItemKeywords.right_click(row, x_offset, y_offset)
+        Clicks.right_click(row, x_offset, y_offset)
 
     @keyword
     def right_click_listview_row_by_index(self, locator, row_index, x_offset=0, y_offset=0):
@@ -289,7 +289,7 @@ class ListViewKeywords(LibraryComponent):
         See `Double Click Listview Row By Index` for details about arguments ``locator`` and ``row_index``.
         """
         row = self._get_row_by_index(locator, row_index)
-        UiItemKeywords.right_click(row, x_offset, y_offset)
+        Clicks.right_click(row, x_offset, y_offset)
 
     @keyword
     def select_listview_cell(self, locator, column_name, row_index):
