+++ src/WhiteLibrary/keywords/items/uiitem.py
@@ -1,8 +1,8 @@
 from TestStack.White.UIItems import UIItem   # noqa: F401
 from WhiteLibrary.keywords.librarycomponent import LibraryComponent
 from WhiteLibrary.keywords.robotlibcore import keyword
-from TestStack.White.UIA import RectX
-from System.Windows import Point, Rect
+from WhiteLibrary.utils.click import Clicks
+
 from TestStack.White.InputDevices import Mouse
 from robot.api import logger
 
@@ -18,7 +18,7 @@ class UiItemKeywords(LibraryComponent):
         mouse position relative to the center of the item.
         """
         item = self.state._get_item_by_locator(locator)
-        UiItemKeywords.click(item, x_offset, y_offset)
+        Clicks.click(item, x_offset, y_offset)
 
     @keyword
     def right_click_item(self, locator, x_offset=0, y_offset=0):
@@ -31,7 +31,7 @@ class UiItemKeywords(LibraryComponent):
         mouse position relative to the center of the item.
         """
         item = self.state._get_item_by_locator(locator)
-        UiItemKeywords.right_click(item, x_offset, y_offset)
+        Clicks.right_click(item, x_offset, y_offset)
 
     @keyword
     def double_click_item(self, locator, x_offset=0, y_offset=0):
@@ -44,7 +44,7 @@ class UiItemKeywords(LibraryComponent):
         mouse position relative to the center of the item.
         """
         item = self.state._get_item_by_locator(locator)
-        UiItemKeywords.double_click(item, x_offset, y_offset)
+        Clicks.double_click(item, x_offset, y_offset)
 
     @keyword
     def get_items(self, locator):
@@ -63,32 +63,3 @@ class UiItemKeywords(LibraryComponent):
         """
         return self.state._get_item_by_locator(locator)
 
-    #Low level function to handle offset click.
-    @staticmethod
-    def click(item, x_offset=0, y_offset=0):
-        offset_position = UiItemKeywords._get_offset_point(item, x_offset, y_offset)
-        Mouse.Instance.Click(offset_position)
-
-    #Low level helper function to handle offset right click.
-    @staticmethod
-    def right_click(item, x_offset=0, y_offset=0):
-        offset_position = UiItemKeywords._get_offset_point(item, x_offset, y_offset)
-        Mouse.Instance.Location = offset_position
-        Mouse.Instance.RightClick()
-
-    #Low level helper function to handle offset right click.
-    @staticmethod
-    def double_click(item, x_offset=0, y_offset=0):
-        offset_position = UiItemKeywords._get_offset_point(item, x_offset, y_offset)
-        Mouse.Instance.DoubleClick(offset_position)
-
-    #Helper function to translate item center to offset point
-    @staticmethod
-    def _get_offset_point(item, x_offset, y_offset):
-        item_bounds = item.Bounds
-        item_center = RectX.Center(item_bounds)
-        offset_point = Point(int(item_center.X) + int(x_offset),
-                                int(item_center.Y) + int(y_offset))
-        if not item_bounds.Contains(offset_point):
-            raise AssertionError("click location out of bounds")
-        return offset_point
